Update Jenkinsfile 250

+++ Jenkinsfile
@@ -188,7 +188,9 @@ deb http://archive.ubuntu.com/ubuntu/ noble-updates main restricted universe mul
 deb http://archive.ubuntu.com/ubuntu/ noble-backports main restricted universe multiverse
 deb http://security.ubuntu.com/ubuntu/ noble-security main restricted universe multiverse
 EOL
-                sudo rm /etc/apt/sources.list.d/*
+                if [ -d /etc/apt/sources.list.d/ ]; then
+                    sudo rm /etc/apt/sources.list.d/*
+                fi
 
                 # Update and upgrade all packages
                 sudo apt-get update
@@ -230,7 +232,7 @@ EOL
 /lib/x86_64-linux-gnu
 /usr/lib
 EOL
-                sudo rm /etc/ld.so.conf.d/*
+                sudo rm -f /etc/ld.so.conf.d/*
 
                 # Rebuild library cache
                 sudo ldconfig -v
